Move meson to be build constant based
