+++ data/be.alexandervanhee.gradia.metainfo.xml.in.in
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <component type="desktop-application">
-	<id>be.alexandervanhee.gradia</id>
+	<id>@APP_ID@</id>
 	<metadata_license>CC0-1.0</metadata_license>
 	<project_license>GPL-3.0-or-later</project_license>
 	<supports>
@@ -10,6 +10,7 @@
 	</supports>
 	<name>Gradia</name>
 	<summary>Make your screenshots ready for all</summary>
+  <launchable type="desktop-id">@APP_ID@.desktop</launchable>
 	<description>
 		<p>Gradia helps you quickly prepare your screenshots for sharing by fixing common issues like transparency and awkward sizing. With simple tools to add backgrounds, adjust aspect ratios, and annotate images, you can enhance your screenshots to look professional and consistent across social media, blogs, and other platforms on the internet.</p>
 		<p>Key features:</p>
@@ -28,9 +29,9 @@
 	<developer id="io.github.alexandervanhee">
 		<name>Alexander Vanhee</name>
 	</developer>
-	<url type="homepage">https://github.com/AlexanderVanhee/Gradia</url>
-	<url type="vcs-browser">https://github.com/AlexanderVanhee/Gradia</url>
-	<url type="bugtracker">https://github.com/AlexanderVanhee/Gradia/issues</url>
+	<url type="homepage">@PROJECT_URL@</url>
+	<url type="vcs-browser">@PROJECT_URL@</url>
+	<url type="bugtracker">@BUGTRACKER_URL@</url>
 	<url type="translate">https://github.com/AlexanderVanhee/Gradia/blob/master/TRANSLATING.md</url>
 	<translation type="gettext">gradia</translation>
 	<launchable type="desktop-id">be.alexandervanhee.gradia.desktop</launchable>
