--- data/be.alexandervanhee.gradia.service.in
@@ -1,3 +1,0 @@
-[D-BUS Service]
-Name=be.alexandervanhee.gradia
-Exec=@bindir@/gradia --gapplication-service
