+++ data/meson.build
@@ -3,30 +3,55 @@ gnome = import('gnome')
 subdir('ui')
 
 desktop_file = i18n.merge_file(
-        input: 'be.alexandervanhee.gradia.desktop.in',
-       output: 'be.alexandervanhee.gradia.desktop',
-         type: 'desktop',
-       po_dir: '../po',
-      install: true,
-  install_dir: get_option('datadir') / 'applications'
+  input: configure_file(
+    input: '@0@.desktop.in.in'.format(PROJECT_RDNN_NAME),
+    output: '@BASENAME@',
+    configuration: conf
+  ),
+  output: '@0@.desktop'.format(APPLICATION_ID),
+  type: 'desktop',
+  po_dir: '../po',
+  install: true,
+  install_dir: join_paths(get_option('datadir'), 'applications')
 )
 
 desktop_utils = find_program('desktop-file-validate', required: false)
 if desktop_utils.found()
-  test('Validate desktop file', desktop_utils, args: [desktop_file])
+  test('Validate desktop file', desktop_utils,
+    args: [desktop_file.full_path()],
+    depends: desktop_file
+  )
 endif
 
-appstream_file = i18n.merge_file(
-        input: 'be.alexandervanhee.gradia.metainfo.xml.in',
-       output: 'be.alexandervanhee.gradia.metainfo.xml',
-       po_dir: '../po',
-      install: true,
-  install_dir: get_option('datadir') / 'metainfo'
+
+# Translate and install Appstream file
+metainfo_file = i18n.merge_file(
+  input: configure_file(
+    input: '@0@.metainfo.xml.in.in'.format(PROJECT_RDNN_NAME),
+    output: '@BASENAME@',
+    configuration: configuration_data({
+      'APP_ID': APPLICATION_ID,
+      'PROJECT_URL': PROJECT_URL,
+      'BUGTRACKER_URL': BUGTRACKER_URL,
+      'HELP_URL': HELP_URL,
+    })
+  ),
+  output: '@0@.metainfo.xml'.format(APPLICATION_ID),
+  po_dir: '../po',
+  install: true,
+  install_dir: join_paths(get_option('datadir'), 'metainfo')
 )
 
-appstreamcli = find_program('appstreamcli', required: false, disabler: true)
-test('Validate appstream file', appstreamcli,
-     args: ['validate', '--no-net', '--explain', appstream_file])
+# Appstream validation
+appstream_cli = find_program('appstreamcli', required: false)
+if appstream_cli.found()
+  test('Validate appstream file', appstream_cli,
+    args: ['validate', '--no-net', '--explain', metainfo_file.full_path()],
+    depends: metainfo_file
+  )
+endif
+
+
 
 install_data('be.alexandervanhee.gradia.gschema.xml',
   install_dir: get_option('datadir') / 'glib-2.0' / 'schemas'
@@ -45,13 +70,12 @@ test('Validate schema file',
      args: ['--strict', '--dry-run', meson.current_source_dir()])
 
 
-service_conf = configuration_data()
-service_conf.set('bindir', get_option('prefix') / get_option('bindir'))
 configure_file(
-  input: 'be.alexandervanhee.gradia.service.in',
-  output: 'be.alexandervanhee.gradia.service',
-  configuration: service_conf,
-  install_dir: get_option('datadir') / 'dbus-1' / 'services'
+  input: '@0@.gschema.xml'.format(PROJECT_RDNN_NAME),
+  output: '@0@.gschema.xml'.format(APPLICATION_ID),
+  configuration: conf,
+  install: true,
+  install_dir: join_paths(get_option('datadir'), 'glib-2.0/schemas')
 )
 
 resources = gnome.compile_resources(
@@ -61,7 +85,7 @@ resources = gnome.compile_resources(
   dependencies: blueprints,
   source_dir: meson.current_build_dir(),
   install: true,
-  install_dir: get_option('datadir') / meson.project_name(),
+  install_dir: PKGDATA_DIR,
 )
 
 subdir('icons')
