+++ gradia/constants.in
@@ -1,0 +1,31 @@
+# Copyright (C) 2025 Alexander Vanhee
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+rootdir          = '@ROOT_PATH@'
+datadir          = '@DATA_DIR@'
+pkgdatadir       = '@PKGDATA_DIR@'
+localedir        = '@LOCALE_DIR@'
+
+app_id           = '@APP_ID@'
+rel_ver          = '@RELEASE_VER@'
+version          = '@VERSION@'
+build_type       = '@BUILD_TYPE@'
+
+project_url      = '@PROJECT_URL@'
+bugtracker_url   = '@BUGTRACKER_URL@'
+help_url         = '@HELP_URL@'
+#translate_url    = '@TRANSLATE_URL@'
