+++ gradia/gradia.in
@@ -29,8 +29,8 @@ import logging
 logging.getLogger("PIL").setLevel(logging.WARNING)
 
 VERSION = '@VERSION@'
-pkgdatadir = '@pkgdatadir@'
-localedir = '@localedir@'
+pkgdatadir = '@PKGDATA_DIR@'
+localedir = '@LOCALE_DIR@'
 
 sys.path.insert(1, pkgdatadir)
 signal.signal(signal.SIGINT, signal.SIG_DFL)
@@ -47,7 +47,6 @@ if __name__ == '__main__':
     gi.require_version('Xdp', '1.0')
 
     from gi.repository import Gio
-
     resource = Gio.Resource.load(
         os.path.join(pkgdatadir, 'gradia.gresource'))
     resource._register()
