+++ gradia/ui/recent_picker.py
@@ -19,7 +19,7 @@ import gi
 gi.require_version('Gtk', '4.0')
 gi.require_version('Adw', '1')
 from gi.repository import Gtk, Adw, GdkPixbuf, Gio, GLib, Gdk
-from gradia.constants import PREDEFINED_GRADIENTS
+from gradia.app_constants import PREDEFINED_GRADIENTS
 from pathlib import Path
 import time
 import re
